Created ConsumeReservationIT cals, fixed ReservationIT
